adds custom hook for Authendication
