+++ src/components/Arbitrations/ArbitrationDetail.js
@@ -1,16 +1,13 @@
-import React, { useEffect, useState } from 'react';
-import {BackButton, Bar, Box, Split, useTheme, Tabs,} from '@aragon/ui';
-
+import React, { useState } from 'react';
+import { BackButton, Bar, Box, Split, Tabs } from '@aragon/ui';
 import { useHistory } from 'react-router-dom';
-
 import DisputeTimeline from './DisputeTimeline';
-
 import ArbDetails from './agreementDetails/arbDetails';
 import AllStatements from './agreementDetails/allStatements';
-import AllProcedure from './agreementDetails/allProcedures';
 import NominationPage from './agreementDetails/NominationPage';
+import useAuthendication from '../../utils/auth';
 import { useAccount } from '../../wallet/Account';
-import wallet from 'wallet-besu';
+
 const networks = require('../../wallet/network');
 
 const NODES = Object.keys(networks).map((node) => {
@@ -18,42 +15,19 @@ const NODES = Object.keys(networks).map((node) => {
 });
 
 const ArbitrationDetail = (props) => {
-  const history = useHistory();
-  const [statementModal, setStatementModal] = useState(true);
-  const openStatement = () => setStatementModal(true);
+  const [tabs, setSelectTabs] = useState(0);
 
+  useAuthendication();
+  const history = useHistory();
   const contractAddress = props.match.params.address;
   const groupId = decodeURIComponent(props.match.params.groupId);
   const walletAccount = useAccount();
-  // Procedure statement modal
-  const [ProcedureStatementModal, setProcedureStatementModal] = useState(false);
-
-  // Procedure Statement form
-  const openProcedureStatement = () => setProcedureStatementModal(true);
-
-  const [tabs, setSelectTabs] = useState(0);
 
   const handleTabChange = (tabs) => {
     setSelectTabs(tabs);
   };
 
-  useEffect(() => {
-    async function load() {
-      try {
-        // Fetching the password locally. Need a secure way to do this for prod
-        const account = await wallet.login(localStorage.getItem('wpassword'));
-        // Update the account context by using a callback function
-        walletAccount.changeAccount({
-          privateKey: account[0],
-          orionPublicKey: localStorage.getItem('orionKey'),
-        });
-      } catch (err) {
-        console.log('ERROR', err);
-        return false;
-      }
-    }
-    load();
-  }, []);
+  useAuthendication();
 
   return (
     <div>
@@ -93,10 +67,10 @@ const ArbitrationDetail = (props) => {
               {tabs === 1 ? (
                 <>
                   <AllStatements
-                      contractAddress={contractAddress}
-                      groupId={groupId}
-                      NODE={NODES[0]}
-                      account={walletAccount.account}
+                    contractAddress={contractAddress}
+                    groupId={groupId}
+                    NODE={NODES[0]}
+                    account={walletAccount.account}
                   />
                   {/*<AllProcedure />*/}
                 </>
@@ -105,10 +79,10 @@ const ArbitrationDetail = (props) => {
               {tabs === 2 ? (
                 <>
                   <NominationPage
-                      contractAddress={contractAddress}
-                      groupId={groupId}
-                      NODE={NODES[0]}
-                      account={walletAccount.account}
+                    contractAddress={contractAddress}
+                    groupId={groupId}
+                    NODE={NODES[0]}
+                    account={walletAccount.account}
                   />
                 </>
               ) : null}
