+++ src/components/Profile/Profile.js
@@ -1,43 +1,11 @@
-import React, { useEffect, useState } from 'react';
-import { useHistory } from 'react-router-dom';
+import React from 'react';
 import ProfileHeader from './ProfileHeader';
-import wallet from 'wallet-besu';
+import useAuthendication from '../../utils/auth';
 
 import { Header } from '@aragon/ui';
 
 function Profile() {
-  const history = useHistory();
-
-  const [isAuth, setIsAuth] = useState(true);
-
-  useEffect(() => {
-    async function load() {
-      try {
-        // Fetching the password locally.
-        const account = await wallet.login(localStorage.getItem('wpassword'));
-
-        console.log('TESTTTT ACC', account);
-
-        const OrianKey = await localStorage.getItem('orionKey');
-        console.log('OIARKEY', OrianKey);
-
-        if (
-          account === null ||
-          account === undefined ||
-          OrianKey === null ||
-          OrianKey === undefined
-        ) {
-          setIsAuth(false);
-          history.push('/login');
-        }
-        console.log('ACCOUNT from Dashboard', account);
-      } catch (err) {
-        console.log(err);
-      }
-    }
-    load();
-  }, [isAuth]);
-  console.log('ISAUTH', isAuth);
+  useAuthendication();
 
   return (
     <React.Fragment>
